Refactored uploading image to it's own function. Problem with new images updating all the old files as well
